pyproject.toml
[project]
name = "singlelitenetplus"
version = "0.1.0"
description = "SingleLiteNetPlus model implementation in PyTorch."
readme = "README.md"
requires-python = ">=3.12"
dependencies = [
    "albumentations>=2.0.8",
    "hydra-core>=1.3.2",
    "matplotlib>=3.10.6",
    "numpy>=2.3.2",
    "opencv-python>=4.11.0.86",
    "pillow>=11.3.0",
    "scikit-image>=0.25.2",
    "scikit-learn>=1.7.1",
    "scipy>=1.16.1",
    "timm>=1.0.19",
    "torch>=2.8.0",
    "torchvision>=0.23.0",
    "tqdm>=4.67.1",
    "wandb>=0.22.2",
]
